Fix bug in function find_programs_in_system
The function was looking for a separator in the PATH variable at a QString
position of QString::length() rather than QString::length()-1. This caused
random abort signals (SIGABRT) due to accessing non-existent memory.

Interestingly this bug has been there since the early stages of development!
However it didn't have a big impact on the logic since accessing a memory
location outside the QString would just give a random value prompting an
insertion of an extra separator on the PATH variable.