Update name of log file in error messages
The configuration log file is now called "configure.log", but the error
messages still referred to the old "config.log". Also changed the C++
configuration script so that messages are only echoed to a separate file
"configure_c++.log" it the c++ configuration is called directly;
otherwise append messages to the main "configure.log".