Fix configuration script for OSX library extension
Get the shared library extension from the egspp_*.conf specifications
file, instead of assuming ".so".

When assessing whether egs++ is installed, the finalize script assumed a
".so" shared object extension, but on OSX it is ".dylib" instead.
Consequently compilation of egs++ codes was skipped on OSX.

In passing, add egs_cbct, egs_chamber and egs_fac to the list of egs++
codes variable, cpp_user_codes.