Fix trailing space in FOUT compilation flag
The current EGSnrc makefiles use no space after the $(FOUT) variable,
because some compilers fail if there is a space between the FOUT switch
and the supplied output name.

But many compilers require a space, so by default the configure script
wrote "FOUT = -o " (note the space) in the configuration file. However,
some editors scrub end-of-line whitespace (which is desirable to avoid
dummy commits) hence it is unsafe to rely on an end-of-line space
character in the script. I updated the script so the space is preserved
in the output, via an $FOUT string variable which includes the trailing
space.