Fix column width in generated machine.macros file
The machine.macros file generated by the configuration script contains
some shell variable expansions that caused the right quotes to spill
over the mortran code width of 80 columns. This raised a compilation
warning. Now the shell variables are expanded in a string and quotes are
properly aligned via a fixed-width printf.