Add ignore patterns for a clean git status
Add ignore patterns so that 'git status' produces a clean output after
the initial EGSnrc configuration. The only untracked files reported
are now the .conf files generated by the configuration script. This also
helps to prevent to some extent the accidental commits of the files
generated by the configure script.