Split density data files in repository
Provide individual compound and element density data files instead
of two concatenated files compounds.del and elements.del that had to be
split during the EGSnrc configuration. With individual data files
directly in the repository, users don't have to reconfigure EGSnrc when
density data files are added or modified. It also makes browsing the
data online easier.

Remove the associated program to split the files, and update the
configuration script to skip the generation of .density files.