Remove overwriting of egs_home in configuration
If the specified egs_home directory exists, the user was asked if they
wanted to overwrite it. Remove this option, because it is dangerous, as
users may loose all changes to their user codes. The script thus now
forces a non-existing directory path for the new egs_home. If the user
wants to remove an old egs_home, they should do it themselves outside
of the script.