Adding egs_beam_install to the system.
This GUI configures EGSnrc and BEAMnrc for an existing system or for
a copy donwloaded from github. It also provides EGSnrc GUIs (egs_gui,
egs_inprz, and egs_view) statically linked to the Qt libraries with
the exception of the Windows version of egs_view. The user environment
is setup to guarantee the system can be used from either the command line,
or from Desktop shortcuts (whenever possible).