Implement magnetic field scaling and Fano source
Implement a Fano source option within the isotropic source. Add
the ability to generate source particles uniformly per unit mass, as
required by the standard Fano test. For this purpose, add the ability to
query the application class for the medium density of any given region.

Implement scaling of the magnetic field strength (B) on a region by
region basis, as required by the second form of the Fano test with a
magnetic field, using a similar scheme as for set_rho_scaling.

There are three options:

1. Scale B in region i by factor rho_i/rho_ref, with rho_ref an input.

2. Scale B in region i by a factor f_i, with f_i an input of the same
   form as medium definition:

   set B scaling = start end f   # 3 inputs
   or
   set B scaling = ir f          # 2 inputs

3. Both rho_ref and a region-based factor

Note that currently the scaling factor is a scalar, hence only the
intensity of the field can be adjusted, but not its direction. This
would require f to be a vector input.