Fix #14: Fix bug in GEO_SHIFT_1 macro in SYNCMLCE
Macro intended to enforce a minimum step of 1e-5 cm provided this is
< USTEP.  Previous version allowed particles to 1) take vanishingly small steps
near leaf boundaries (infinite loops) and 2) take a step > USTEP.  Definition
of macro now matches that in MLCE, which was corrected in 2009.  Propagation
of this bug must be due to the fact that SYNCMLCE was developed from an
older version of MLCE.  (Bug found by user Michael Martyn at National University
of Ireland, Galway).