Improve magnetic field scaling and the Fano source
Update the isotropic source to include the Fano geometry part provided
by Hugo Bouchard. Move the rejection sampling block inside the geometry
sampling block, to avoid costly direction sampling for events that are
ultimately rejected. Also, use a multiplication instead of a division
for sampling the Fano source using an isotropic source and the
IncludeAll directive.

Fix the value of the magnetic field strength B in the current region.