Fix egs_view region display overflow bug
On a ~1300 region file of mine, sometimes the region list in the
display would overflow. It appeared that the "draw regions" while
loop was continuing, even though reg > maxreg. Changing form to
a for loop fixed the bug. The logic appeared sound, so I expect
something was trampling on memory.

In the process I corrected a `delete` to `delete[]` to comply
with Address Sanitizer.