Fix egs_view thread porting regression where simple scenes black out
The change covered the else case of a branch in the dynamic render scaling
section. This branch is taken when it takes less that 30ms to render a frame at
full resolution; before this patch, when that happened, the requested view
scaling was equal to the view size so that only the top-leftmost pixel was
shown on screen.

Included are changes that prevent the ImageWindow from rendering a scene
before `render(...)` has been called at least once with an actual Geometry.