Fix #23: Change file names for case-insensitive OS
Some file systems are case-insensitive. For example OS X relies on
HFS by default, which is case-insensitive. Hence when EGSnrc is cloned
to a Mac, changes are automatically applied to the repository because of
duplicate image file names (I6702 and i6702). Although not critical, it
is annoying. Change file names to avoid conflicts.