Implement new code style standard in egs++ library
This is a format-only commit. The egs++ source code before and after
this commit compiles to identical object files.

The coding style is not consistent throughout the egs++ library. Use the
"Artistic Style" (astyle) code formatter to clean up the egs++ source
code: make indentation uniform at 4 spaces, split compound statements,
adjust braces and parenthesis placement and spacing, and indent
preprocessor code blocks.

It is better to do this in one swoop so as to avoid a large number of
format-only commits polluting the code history later on.

Add scripts/egs_style_standard file which contains the astyle options
to apply to new code, or after editing existing code:

astyle --options=$HEN_HOUSE/scripts/egs_style_standard "*.cpp" "*.h"

For reference, the options used for astyle 2.05 for this commit were:

--convert-tabs
--indent=spaces=4
--style=attach
--unpad-paren
--pad-header
--align-pointer=name
--break-closing-brackets
--add-brackets
--indent-preproc-block