Fix script to handle missing trailing spaces
The C++ configuration script sed commands tacitly relied on a trailing
space in empty variables in the EGS_CONFIG file. Add a "\?" switch in
sed to match cases where the variable is not defined; otherwise one ends
up working with the variable name instead of its value.

This is the root cause of issue #20.