Move painting work offthread in egs_view
Now the RenderWorker generates, scales, and adds axes labels
to the image that is requested of it. This makes it easier to repaint
the image window, especially when nothing is happening, and also
lets us save exactly what is on screen (skipping the region markers).

The scaling routine is self-written because the QT scaling routines
are too general and thus too slow.

And, in the fallout, modify several other things:
- Fix a stalling bug when you zoom in too far on the main axes
- Modify saving code to ensure that you can't try to save an
  image while another is in progress, by disabling save button