Look for density correction files in EGS_HOME
Density correction files were only sought inside the "compounds" and
"elements" directories under $HEN_HOUSE/pegs4/density_corrections.
Beyond a message in the .mederr file, no alert was raised if the
requested file was not found.

Now the directory $EGS_HOME/pegs4/density is scanned in priority for
user-provided density correction files. If the file can't be opened the
program stops with an error message to that effect.