Add a source for Fano testing with magnetic fields
Implement the Fano source in its own right (instead of an option in the
isotropic source class). The source relies on a rejection scheme to
generate source particles uniformly per unit mass, which is a condition
of the standard Fano test. It also generates particles isotropically,
which is one possible additional condition for proper Fano testing in
the presence of a magnetic field. The possibility to generate particle
non-isotropically should be added.