Fix pulse height scoring in tutor7pp
Events depositing the maximum particle energy are excluded from the
histogram. As a consequence, large differences are observed in those
cases if one compares the energy deposited in a given region calculated
from the fraction of energy deposited to the sum over all histogram
bins. Now when Emax is deposited in a region, it is scored in the last
bin, i.e., the upper bound of the last bin is closed.