Improve radiative Compton warning message
When turning on radiative Compton corrections one needs to include
file rad_compton1.mortran as part of the source files and set radc_flag
to 1 via input or on the user's application. If one forgets to add
the above file, a warning is issued and radc_flag is set to 0.

- Fix typo in warning message

  The warning message was referring to rad_compton.mortran instead of
  rad_compton1.mortran.

- Set $RADC_HATCH to call $RADC_WARNING by default

  Later on if the user includes radc_compton1.mortran in the building
  process, this macro is replaced by a call to subroutine radc_init.
  This way, if the user forgets to include the above file, radc is
  turned OFF and a warning is issued.