Fix uninitialized values/compiler errors in cones
This patch resolves a number of compiler warnings for the egs_cones
geometry.  Most importantly is the change in EGS_ConeSet::hownear to
properly initialize the tc variable. If hownear is called with ireg=0 or
ireg=2*nc then `tc` never gets set and the tco < tc comparison in the
last line of the function will produce undetermined results.