Fix #67: Fix density correction file search paths
* Fix #67: Fix inconsistent search paths for density correction file.

Consistent search paths and order for density correction files between
pegs4 and codes running in pegsless mode.  Also, changed egs_gui so
that it always calls pegs4 with full density correction file name.
Finally, changed egs_gui so that, on executing pegs4, if the density
correction file is not found, an error dialog pops up.  Order of
search for density correction file is now:

1. full pathname
2. $EGS_HOME/pegs4/density_corrections/
3. $EGS_HOME/pegs4/density_corrections/elements
4. $EGS_HOME/pegs4/density_corrections/compounds
5. $EGS_HOME/pegs4/density/ -- for pegsless to preserve compatibility
with past installations where this directory was created
6. $HEN_HOUSE/pegs4/density_corrections/elements
7. $HEN_HOUSE/pegs4/density_corrections/compounds


* Fix #67 (related): Fix general bugs in egs_gui

Fix a couple of bugs when running pegs4 from egs_gui:

1. Segmentation fault when reading in medium comp using
   density correction file and then running with
   ICRU density corrections switched off.
   composition_table was looking for undefined
   items.

2. Now outputs an error if user must explicitly
   define medium as a gas.