Allow negative z-axis vector as a rotation vector (#107)
The description of the 'rotation vector' input of affine transforms
specifies that it "defines a rotation which, when applied to the 3D
vector defined by the input, transforms it into a vector along the
positive z-axis". Thus, it's natural to input '0, 0, -1' in order to 
rotate an object 180 degrees along the z-axis.