Improve CD geometry's howfar (#87)
* Improve CD geometry's howfar

Two cases are mishandled by the howfar method:

1) A particle is approaching the geometry, but due to round off, ends
up sitting exactly on the boundary. At that point, it will be considered
inside the CD geometry which causes a mismatch with its global region
number (0). The problem is that the check for condition (b) in the code
fails to catch this case. When the particle sits on the boundary, the
howfar call will set 'tb' to the distance to the next boundary, which
can be much larger than 'epsilon'. Solution: call howfar, but in the
opposite direction, which should set 'tb' to 0 if sitting on the
boundary.

2) A particle has just left the geometry but is still sitting exactly
on the boundary. This corresponds to condition (a) in the code. The
problem arises with concave geometries. If the particle is sitting on
the boundary and pointing away from the CD geometry, the howfar call
will set 'tb' to 0. But for CD geometries made of concave geometries,
the particle may re-enter the CD geometry on its next step, so we should
instead go to checks 1. and 2.

* Fix isConvex() check