Fix #152: input bug in beamdp/beamdp_gui option 11
Allow the specification of the range of particles for which to list
parameters, as required by beamdp. Previously, beamdp_gui only allowed
an input for the total number of particles for which to list parameters,
instead of the range. Confusion was exacerbated by the command-line
description of this input from beamdp itself, which also referred to the
number of particles instead of the range. This has been fixed as well.

Also fix a serious bug resulting in erroneous data output for all
options when analyzing IAEA phase space data.