Fix typo in CODATA electron rest mass value
Adjust the electron rest mass to 0.5109989461 from the previously
entered value 0.510998946131. The last two digits "31" in the latter are
in fact the uncertainty on the last two digits "61".

Also add a reference to CODATA as the source of the electron rest mass
value, change a couple "/1000" to "*0.001" and fix a small residual bug
in iaea_phsp_macros.