Add epsilon global constant for float comparisons
Add a new const variable 'epsilon' in egs_functions to be used instead
of hard-coded small values when comparing floating-point numbers. The
constant 'epsilon' can be used anywhere in EGSnrc. By contrast, the
boundaryTolerance variable is used only in geometry howfar() functions
and can be set in the input file.