Fix several bugs in radionuclide source
Fix a bug that occured when all gammas for a given excited daughter
state had zero probability of emission.

Fix a bug with electron capture where returning a zero energy particle
threw off the particle count. Now getNextParticle() loops until it gets
a particle with nonzero energy.

Fix some floating point errors.

Fix a bug where all x-rays had an energy 10 times too large.

Remove the branch multiplier for ENSDF normalization records. I'm not
sure if this is correct in general, but it works for the ENSDF files
provided by LNHB, for now.

Check for the existence of ENSDF files in order to exit gracefully.

Use the 'charge' input correctly.

Improve the sampling efficiency of getNextParticle().

Fix a bug in the fluence, which was off by 1.