Disable GASP when using a density correction file
Disable GASP (gas pressure) density corrections when a density
correction file is specified. If both a density correction file and a
GASP>0 are specified, the value of GASP is reset to 0.

Previously, the pegsless implementation confusingly crashed if both were
specified and GASP was different from 1.

Accordingly, disable the gas pressure input in beamnrc and dosxyznrc
guis when a density correction file is specified.