Fix radionuclide ENSDF x-ray intensities
Fix a problem with the energies of x-rays from radionuclide sources
using ENSDF files. In the cases where x-ray data is provided in ENSDF
comments with a line denoting the 'total' contribution from a series of
lines, don't rely on the total. Instead, assign each line its own
probability. If the intensities are provided, use them. Otherwise,
assign each line an equal fraction of the intensity from the 'total'.

This change avoids a problem where a single x-ray line (total) could
have an artificially large intensity, instead of several lower intensity
lines. It also avoids having an x-ray line with a non-physical energy,
since the 'total' lines may be an average of several lines.