Add option to skip DOSXYZnrc source 20 calibration
Add a flag to optionally skip the calibration run performed to refine
NRCYCL when using source 20 in DOSXYZnrc. If the flag is set to 1, the
calibration run is skipped. To ensure backwards compatibility, the
default value for the flag is 0, and the calibration always runs if the
flag is omitted in the input file.