Recognize compiler combinations in egs_configure
Recognize when the Fortran or C++ compiler changes in the gui and
updates the compiler arguments accordingly. Previously the gui was not
updating the Fortran libraries needed to link Fortran and C++ source
code.

Also use "make -j" to build the egs++ library, speeding up the
compilation significantly on multicore machines.