Fix #179: electr bug when ustep is zero
Fix a bug that happens rarely when the random number rnnoss is set to
zero. This results in ustep=tustep=0 but occasionally they are set to
an erroneous value. With electric and magnetic fields this leads to
infinite loops.

Now, when the random number is less than 1e-30 it is set to 1e-30.