Fix #100: egs_gui bug with density file check box
Fix egs_gui bug with the density correction file check box. Previously,
when unchecking and re-checking the ICRU density correction box, PEGS4
ran with the density correction file but with medium type specified as
"Compound" (and potentially with a different composition than that
specified in the density correction file). In this case, PEGS4
confusingly failed.