Replace hard coded values with veryFar constant
Clarify the code by using a variable explicitly named veryFar instead of
hard coded values for large distances (leave the rhomin = 1e30 in
egs_ndgeometry hard coded for now since it is not a distance).

There are four files in which the hard coded constant was different from
1e30. It is not immediately clear in those cases whether a different
constant was intentionally used to avoid edge cases in floating-point
comparisons:

  HEN_HOUSE/egs++/geometry/egs_nd_geometry/egs_nd_geometry.h
  HEN_HOUSE/egs++/geometry/egs_planes/egs_planes.h
  HEN_HOUSE/egs++/geometry/egs_vhp_geometry/egs_vhp_geometry.h
  HEN_HOUSE/egs++/geometry/egs_spheres/egs_spheres.cpp

To avoid unforeseen consequences, the veryFar constant is in those cases
multiplied by a hard coded constant to preserve the original value. This
should be cleaned up eventually.