Fix sub-threshold fluorescence and auger auscalls
Fix a bug resulting from sub-threshold fluorescence and auger auscalls
that did not add a particle to the stack. Some codes depended on the top
stack particle being from a fluorescence of auger event, but in the
sub-threshold case this was not true. Now there are new sub-threshold
auscalls that are independent, and used for the kerma calculations in g
and dosrznrc.