Initial implementation of Travis-CI automation
Add a simple travis configuration file to automate testing software
builds on each pull request. Travis-CI is a continuous integration
service that automates building and testing software. As a starting
point, the travis script simply clones the repository branch and
attempts to configure EGSnrc with default parameters, to trap basic
configuration failures.

Add a configure.expect script to answer the EGSnrc configuration script
questions automatically. This requires the expect utility to be
available.

Adjust the egsnrc_bashrc_additions to avoid returning an error when
optional local additions files don't exist.

Ensure the travis check fails when any part of the configuration fails.
Some critical failures cause the configuration to abort with status 1
(error), but most failures merely echo a "failed" status to the
terminal, hence travis does not trap those because it reports failure
upon a non-zero return value. As a quick solution, tee the configuration
terminal output message to file and use grep to trigger a status of 1
if the file contains "fail":

! grep -i fail configure.log