Fix LATCH and ZLAST in DBS when z(np)>ssd (#247)
Use a GOTO jump to the default LATCH (and ZLAST) setting routine in
AUSGAB instead of immediately returning when z(np) > ssd.  Previously,
no further LATCH bits were set once z(np) was greater than the ssd of
the splitting field, and ZLAST for these particles was not set.

This is probably a low incidence bug, given that in general users set an
ssd which corresponds to the bottom of the accelerator.