Check for nfields > maxfield in BEAMnrc sync CMs
Add a check in the BEAMnrc synchronized (SYNC) component modules (CMs)
to make sure that the number of fields is less than the allowed maximum,
as defined in the macros file. Inform the user about the problem and how
to resolve it. Previously, the program simply crashed with a
segmentation fault when NFIELDS > MAXFIELD.