Fix #182: increase maxfield for BEAMnrc sync CMs
Increase the MAXFIELD variable for the BEAMnrc synchronized (sync)
component modules (CMs). It is common to use these CMs for a large
number of fields (over 10000 for tomotherapy), but 1024 is a reasonable
default maximum in general.

Add a check in the BEAMnrc sync CMs to make sure that the number of 
fields is less than the allowed maximum, as defined in the macros file. 
Inform the user about the problem and how to resolve it. Previously, the 
program simply crashed with a segmentation fault when NFIELDS was 
greater than MAXFIELD.