Fix #260: rest mass not initialized in C interface
Fix the rest mass initialization for EGSnrc applications that depend on
the C interface. The bug fix in the parent commit moved the declaration
of the rest mass inside the COMIN/USEFUL/ macro. However this in turn
breaks C and C++ EGSnrc applications because the C interface relies on
its own declaration of the COMIN/USEFUL/ macro, which also has to be
updated with the rest mass data statement.