Invoke relax with actual shell number from COMPT
Invoke relax using the actual shell number j, if $EADL_RELAX is true.
Previously, the COMPT subroutine was calling relax with shn_array(j)
which is an integer between 1 and 7 to indicate whether the shell is K,
L1-L3, <M>, <N>, or beyond the N shell.

Add a comment in subroutine PHOTO to note that if $EADL_RELAX is true,
k=5 creates a vacancy for M1 rather than <M>.

Add a comment in subroutine eii_sample to note that relax is invoked for
K, L1, L2, and L3 shells only.