Add radionuclide edep contributions
Add the ability for the radionuclide source to contribute to dose within
the source region. This requires some additions to the application class
and to the dose scoring object (to allow a region as argument). The
region of the top particle on the stack cannot be used because there may
be no particle on the stack in the first call to getNextParticle().

The dose scoring for the radionuclide source is required to account for
sub-threshold contributions from relaxations and local absorption of
alpha emissions.