Fix charged particle fluence scoring in cavity
Although PIRS-898 states that cavity can calculate the e-/e+ fluence,
this is not entirely correct. The quantity that is actually calculated
seems to be something like the product of the fluence times the stopping
power.

Now the actual charged particle fluence is calculated using two methods:

- Algorithm already in place in cavity, but dividing by the stopping
  power for the lower edge energy L(E_i) of bin i.

- Algorithm used in FLURZnrc for comparison.

The first one is more accurate as it does not assume that the stopping
power is constant along the condensed history step, as opposed to the
FLURZnrc algorithm.

By default the first algorithm is used. To switch to a FLURZnrc-like
calculation of the fluence, use the following input key in the fluence
scoring input block:

:start scoring options:
    ...
    :start fluence scoring:
        ...
        method = flurz # stpwr (default) or flurz
        ...
    :stop fluence scoring:
    ...
:stop scoring options:

Range rejection is disabled for a FLURZnrc-like calculation of the
fluence and this calculation type relies on the knowledge of the
particle step size. For the stpwr calculation type (using stopping
powers), a warning is issued because a proper selection of ESAVE is
required for accurate results. As a general rule, avoid using range
rejection when calculating charged particle fluence!