Fix atomic relaxation vacancy dependency on ecut
Check only against the RELAX-CUTOFF variable for energy cutoff on
vacancy creation in both relaxation routines, in order to make the
calculation of g independent of the cutoff energy ecut. Also fix a bug
in the output of the g application.

Previously, the energy cutoff for vacancy creation was the maximum of
ecut and pcut, or the RELAX-CUTOFF. Thus changing ecut caused a
different chain of relaxations, hence results related to sub-threshold
contributions (e.g., the value of g) were ecut dependent.