Fix photo-electron emission when relaxation is on
Emit a photo-electron with the photon energy (instead of depositing the
photon energy locally) when $EADL_RELAX = true and the photon interacts
with a shell below L3.

Likewise, emit a photo-electron (instead of depositing energy locally)
when $EADL_RELAX = true and the photon energy falls below the binding
energy of the M2 shell, i.e., binding_energies($MXSHELL,iZ).