Fix AUSCALL for relax vacancy cut-offs
Fix the AUSCALL related to relaxation vacancies below the cut-off
energy, or in the outer shell. The AUSCALL was changed to always treat
the energy deposition as though it came from an electron. Previously, it
was differentiated based on the charge of the particle initiating the
relaxation. All references to relax_initiator_q have been removed, since
they are no longer used.

This bug was discovered by plotting mu_en/rho as a function of energy,
using the g application. The value of mu_en/rho for air was
underestimated, due to the assignment of energy to radiative losses in
these vacancy cut-off depositions.

This bug only affected applications using iarg=33 and 34. Namely, g and
dosrznrc. Additionally, this bug never made it into the master branch.