Add .3ddose output to egs++ dose scoring object
Add an option to the EGS_DoseScoring ausgab object to output doses to a
file in .3ddose format (i.e. standard format for DOSXYZnrc output) for
any EGS_XYZGeometry instance in the simulation geometry. See details in
egs_dose_scoring.h.

This is accompanied by a small change to the EGS_AdvancedApplication
finishSimulation() method. At the end of a parallel run, i_parallel is
now set to 0, so we can use EGS_Application::getIparallel() to query
whether or not this is the last job and, only then, to output a .3ddose
file containing the combined results.