Rename ibsearch in DOSXYZnrc, remove from EDKnrc
Rename the binary search function in DOSXYZnrc to avoid conflict with
the new binary search function in egs_utilities.mortran (needed by the
new photoelectric routine). The function name is changed from ibsearch
to ibsearch4, reflecting the fact that it takes a real*4 array.

Remove the binary search function ibsearch from the EDKnrc application
source code. This function is now defined in egs_utilities.mortran, and
in fact it is not even used in EDKnrc.