Fix #318: changing MXSTACK in array_sizes.h
Add array_sizes.h to the compile dependencies for egs_interface2.c in
egspp1.spec. As discovered by @ojalaj when running egs_chamber, without
this dependency, changes to MXSTACK (the maximum dimension of variables
in the STACK Fortran common block) were not propagated through to the
EGS_Stack definition in egs_interface2. This resulted in mismatched
variable dimensions and null results.

A workaround was to run "make clean" before "make" after changing
MXSTACK, but this forced a recompilation of egs_interface2.c.