Adjust water and graphite density correction files
Update the water_icru90 density-effect correction data to match the
density of 0.998 g/cm3 in ICRU Report 90, Table A.3, page 85. The
density of water is 1 g/cm3 at 0C, but 0.998 g/cm3 at 20C. Add the ESTAR
calculation parameters in the first line comment, and change the file
name to lowercase icru.

Add the ESTAR calculation parameters in the first line comment in the
graphite density-effect correction files. These parameters follow the
ICRU Report 90 recommendations. Change the file name to lowercase icru.

Note that the delta values for graphite differ from those published in
Report 90 (Table A.2, page 84). But the differences are only significant
at low energies, where delta itself is small. The difference is due to
considering only a single conduction electron in graphite, instead of
two in the ESTAR program. This is explained in ICRU Report 90, page 26.