Fix #349: fix pegsless mode in DOSXYZnrc source 20
Pass on the pegsless mode information to the BEAMnrc shared library when
using source 20 in DOSXYZnrc, so that the BEAMnrc simulation also runs
pegsless: don't forget to define the media in the BEAM input file!

With source 20, it is assumed that the pegs data is the same for the
BEAMnrc and DOSXYZnrc parts of the simulation. Previously, the pegsless
mode information was not properly relayed to the BEAMnrc shared library.