Fix egs_view alias to accept input argument
Remove subshell parentheses from the egs_view alias, which prevent it
from accepting input arguments. The subshell was used to prevent
overriding the user's LD_LIBRARY_PATH environment variable, but this is
not necessary since the variable scope here is strictly limited to the
egs_view executable (by omitting the semi-colon after the variable
definition, it becomes part of the executable invocation command).