Remove outdated, meaningless version numbers
Delete all references to version or revision numbers dating back to when
the source code was under CVS control (or even earlier, under RCS), or
replace them by empty strings.

Currently version is controlled by git for the source tree as a whole.
It is referenced by the release year, or else by the commit hash value.