Fix radiative splitting csplit initialization
Initialize the csplit variable to 1 by default in the egs_chamber
application constructor; csplit is used to read in the radiative
splitting number from the input file.

Previously, csplit was only initialized if the input contained a
variance reduction input block. This caused a serious bug to be
introduced in commit 50e0d33, whereby the_egsvr->nbr_split was set to
the arbitrary (uninitialized) csplit number in the shower loop when
there was no variance reduction input block. If this number happened to
be a large positive integer (about half the time), any radiative event
immediately exceeded the stack size and aborted the simulation.

This probably went unnoticed for so long because one rarely runs
egs_chamber without variance reduction.