Add target 'realclean' in egs++ makefile
Add a 'realclean' phony target which removes all objects files and
libraries from the dso/config folder that belong to the egs++ library
including sources and geometries. Useful when one wants to recompile the
full library.

The 'clean' target only removes object files in the egs++ folder without
cleaning objects for geometries, sources, etc. It is meant to only
rebuild the basic part of egs++.

Note that libiaea_phsp.so (iaea_phsp.lib on Windows) is preserved.

Fix the 'realclean' target for guis as well, to make it general for
Linux and Windows and use the proper Makefile name.

Also expand the 'clean' rule in the iaea_phsp makefile to remove all
object files.