Add egs_view config file and update layout
Add the ability to save and load egs_view configurations, including
material colors, clipping planes, view orientation, etc. The
configuration file can be passed as an argument or loaded in the gui.

Change the layout of the viewer to hide away options that are not used
as often. Tabs have been added to support further development.