Add tools to create EGSnrc distribution files
Currently only compressed and uncompressed tar archives are created for
the pre-compiled guis (egs_gui, egs_inprz, and egs_view). The latter
requires the geometry, shape and track scoring modules to be included.
The guis are linked statically with the Qt library for distribution.
However it could also be used with a dynamic Qt library as long as the
targeted systems use the same Qt and OS libraries.

The egs_configure gui is also statically built, and compressed
*.egssself archives are appended to it for self-extraction.

Build process on Linux is done directly, BUT in Windows one must have
first built the guis. To do this, one must run make (or mingw32-make if
using MinGW) in $HEN_HOUSE/pieces/windows. Once the Windows part is
completed, running make in $HEN_HOUSE/distribution finishes the build
process.

The egs_guis_list file specifies the name of the archive file and files
to be included, in a special format which is described in the header of
egs_guis_list. Currently only the guis and the egs++ geometry are
included, but one could create arbitrary archives, including the
documentation for example.