Add predefined data for SSB media in pegsless mode
Add precalculated data for Sternheimer-Seltzer-Berger (SSB) media in the
pegsless implementation, to be consistent with pegs4.

Precalculated density effect data available for a small subset of media,
referred to in pegs4 as SSB media, were not included in pegsless, even
though the lookup table for the names of these media was. This caused a
crash in the (rare) pegsless run in which an SSB medium was referred to
by its exact name (e.g., "H2O" instead of "h2o" or "H2OICRU512", etc.)
and a density correction file was not specified.