Fix tutor7pp pulse-height distribution regions
Fix a bug in tutor7pp where the pulse-height distribution (PHD) regions
specified in the input file were not associated with the correct
regions. The PHDs that were output corresponded to ir+1, but the input
and output suggested that they corresponded to ir. Now, the PHDs are
output for the correct regions and there is additional description of
the output format.