Add egs_view camera position and bug fixes
Add camera position controls, and enforce the rotation point to be
inside the geometry. Add additional hunting for an initial point inside
the geometry so that geometries far from the origin still have a chance
of being found.

Fix a bug where the axis flipped direction when the geometry was not
near the origin.

Improve the region list display and the scaling that occurs with window
resizing.

Fix the clearGeometries() function in egs_base_geometry so that it
removes composite geometries properly. This fixes a crash that occurred
when reloading inputs.

Add a QT version check before requesting a monospace font style hint.