Fix #388: concave cd geometry errors
Address a special case where convex geometries result in a concave cd
geometry when combined. This showed up as geometry errors and could be
created using a radius-mismatched cylinder and sphere in a cd geometry.
Also fix a bug where a small step is taken in the cylinder or sphere
howfar() routine by boundaryTolerance/2 instead of boundaryTolerance,
causing checks against the boundaryTolerance to fail in the cd geometry.