Add region exit and entry pairs in phsp scoring
Add an option to score particles in the phase space when it goes from
one specific region to another.

Add the functionality to include or exclude multiple crossers (EGSnrc
format phase space data only). This requires a new method setLatch in
EGS_AdvancedApplication (and the equivalent virtual function in
EGS_Application) to set latch bit 31 once a particle has been scored.

Move the description to the header file and apply astyle coding standard
to the new egs_phsp_scoring files.