Set defaults for global ECUT and PCUT in guis
Save the global ECUT and PCUT values as "0" in the .egsinp file if they
are left blank in the guis.

This results in the following default behaviour in subroutine HATCH:

1) For BEAMnrc, region-by-region setting of ECUT(i) or PCUT(i) to
   AE(med(i)) or AP(med(i)), respectively.

2) For DOSXYZnrc, setting of the single value of ECUT or PCUT that
   applies to all regions to the maximum of value of AE or AP over all
   media.

Modifications to BEAMnrc and DOSXYZnrc were necessary, to ensure that
ECUTIN or PCUTIN, the actual input variable for global ECUT or PCUT, was
also set to a value other than zero. ECUTIN, specifically, is used to
determine range for range rejection and running with this set to zero
may cause a crash.

In BEAMnrc, this means resetting ECUTIN or PCUTIN to AE or AP of the
default medium (or the maximum AE or AP over all media if the default
medium is vacuum). For DOSXYZnrc, ECUTIN or PCUTIN gets reset to the
maximum AE or AP over all media (i.e., the new value of ECUT or PCUT set
in HATCH).