Fix egs_view sim geom selection and minor bugs
Fix the egs_view simulation geometry selection so that only geometries
that are actually named in the input file can be selected. When
selecting the simulation geometry, the view window now resets the camera
position to find a point within the new geometry.

Fix a bug where the axis did not display correctly at certain view
angles when the geometry was not near the origin.

Fix a bug in egs_cylinders where the geometry type was not defined. This
is necessary so that the type displays correctly in egs_view.